requirements.txt
Flask>=3.1,<4.0
Flask-SQLAlchemy>=3.1,<4.0
requests>=2.32,<3.0
gunicorn>=23.0,<24.0
